8228400: Remove built-in AArch64 simulator Reviewed-by: adinn, aph, dsamersoff
